requirements.txt
Flask
Flask-SQLAlchemy
Flask-Admin
Flask-Security
bcrypt
Flask-Migrate
Pillow